# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kleines Mypage API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get active users",
                "description": "Get active users with emails and birthdays for `MASTER` and `MANAGER`. Otherwise, emails and birthdays are hidden",
                "operationId": "3a413512c03716149b2218690364c1e0",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/User"
                                                    },
                                                    {
                                                        "properties": {
                                                            "profile": {
                                                                "allOf": [
                                                                    {
                                                                        "$ref": "#/components/schemas/Profile"
                                                                    },
                                                                    {
                                                                        "properties": {
                                                                            "birthday": {
                                                                                "type": "string",
                                                                                "example": "2000-01-01"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                ]
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    {
                                                        "properties": {
                                                            "email": {
                                                                "type": "string",
                                                                "example": "admin@chorkleines.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            },
                                            "property": null
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 401
                                                },
                                                "detail": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Forbidden"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 403
                                                },
                                                "detail": {
                                                    "type": "string",
                                                    "example": "Forbidden"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profile": {
                "properties": {
                    "last_name": {
                        "type": "string",
                        "example": "山田"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "太郎"
                    },
                    "name_kana": {
                        "type": "string",
                        "example": "ヤマダタロウ"
                    },
                    "grade": {
                        "type": "integer",
                        "example": 18
                    },
                    "part": {
                        "type": "string",
                        "example": "T"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "PRESENT"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "MASTER"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}